refactor: moved openai setup and functions to utils
